댓글 수정 삭제 로직 리팩토링

+++ src/actions/studyroom/commentActions.ts
@@ -6,7 +6,7 @@ import { revalidatePath } from 'next/cache';
 
 export async function getComments(targetId: string) {
   const supabase = supabaseServer();
-  const {data, error} = await supabase.from('comments').select('*, user(name, images(url)), reactions(*)').eq('target_id', targetId).order('created_at', {ascending: false});
+  const {data, error} = await supabase.from('comments').select('*, user(name, images(url)), reactions(*)').eq('target_id', targetId);
 
   if (error) {
     return null;
@@ -15,7 +15,7 @@ export async function getComments(targetId: string) {
   return data;
 }
 
-export async function upsertComment(targetId: string, comment: string) {
+export async function insertComment(targetId: string, comment: string) {
   const supabase = supabaseServer();
   const { data: { user } } = await supabase.auth.getUser();
   const userId = user?.id;
@@ -26,7 +26,7 @@ export async function upsertComment(targetId: string, comment: string) {
 
     const { data, error } = await supabase
     .from('comment')
-    .upsert({ user_id: userId, target_id: targetId, comment: comment })
+    .insert({ user_id: userId, target_id: targetId, comment: comment, created_at: new Date() })
     .select()
     if (error) {
       throw new Error(`Insert Comment has an error. ${error.message}`);
@@ -35,6 +35,31 @@ export async function upsertComment(targetId: string, comment: string) {
     return data;
   } catch (err) {
     return err;
+  }
+}
+export async function updateComment(id: number, comment: string) {
+  const supabase = supabaseServer();
+  const { data: { user } } = await supabase.auth.getUser();
+  const userId = user?.id;
+
+  try {
+    if (!userId) {
+      throw new Error('There is no User');
+    }
+
+    const { data, error } = await supabase
+    .from('comment')
+    .update({ comment: comment })
+    .eq('id', id)
+    .eq('user_id', userId)
+    .select()
+    if (error) {
+      throw new Error(`Insert Comment has an error. ${error.message}`);
+    }
+    revalidatePath('/');
+    return data;
+  } catch (err) {
+    return err.message;
   }
 }
 
@@ -97,53 +122,24 @@ export async function createReaction({targetId, emoji}: {targetId: string, emoji
   }
 }
 
-export async function updateComment({id, comment}: {id: string, comment: string}) {
+export async function deleteComment(id: number) {
   const supabase = supabaseServer();
-  const userId = await getServerUserId();
-
+  const { data: { user } } = await supabase.auth.getUser();
+  const userId = user?.id;
   try {
-    if (!userId) {
-      throw new Error('There is no User');
-    }
-    if (comment === '') {
-      throw new Error('comment is required');
-    }
-    const { data, error } = await supabase
-      .from('comments')
-      .update({
-        comment,
-      })
-      .eq('id', id);
-
-    if (error) {
-      throw new Error(`${error}`);
-    }
-    return { success: true, data };
-  } catch (err: any) {
-    // TODO 에러 타입
-    return { success: false, error: err.message };
-  }
-}
-
-export async function deleteComment({id}: {id: string}) {
-  const supabase = supabaseServer();
-  const userId = await getServerUserId();
-  try {
-    if (!userId) {
-      throw new Error('There is no User');
-    }
-    if (id === '') {
-      throw new Error('comment id is required');
+    if (!userId || !id) {
+      throw new Error('There is an error');
     }
     const { error } = await supabase
-      .from('comments')
+      .from('comment')
       .delete()
       .eq('id', id);
 
     if (error) {
-      throw new Error(`${error.message}`);
+      throw error;
     }
-    return { success: true };
+    revalidatePath('/');
+    return {success: true}
   } catch (err: any) {
     return { success: false, error: err.message };
   }
